Add zip traversal complete solution

+++ incf/preprocess/zip_traversal.py
@@ -13,40 +13,61 @@ def extract_zip(path):
     # get the suffix (ses-preop or ses-postop)
     suffix = get_suffix(path)
 
-    # open zip file
-    archive = zipfile.ZipFile(path)
+    #
+    contents = os.listdir(parent)
 
-    # store all newly added files
-    added = []
+    if [f'weights_{suffix}.txt', f'distances_{suffix}.txt', f'centres_{suffix}.txt'] in contents:
+        return
 
-    # iterate over zip content and extract only the files that are included in TO_EXRACT
-    for ext in conv.TO_EXTRACT:
-        if ext in archive.namelist():
-            # add the suffix to the newly extracted file
-            new_filename = os.path.join(parent, add_suffix(ext, suffix))
+    # rename existing files and skip zip file
+    for file in contents:
+        if file.endswith('.zip'): continue
 
-            # rename tract_lengths.txt to distances.txt
-            if 'tract_lengths' in new_filename:
-                new_filename = new_filename.replace('tract_lengths', 'distances')
+        basename = os.path.basename(file)
 
-            if not os.path.exists(new_filename):
-                # extract new file
-                archive.extract(ext, path=parent)
+        if suffix not in basename and not basename.endswith('txt'):
+            new_file = os.path.join(parent, file.replace('.', f'_{suffix}.'))
+            os.replace(os.path.join(parent, file), new_file)
 
-                # rename the new file
-                os.replace(os.path.join(parent, ext), new_filename)
+    if f'weights_{suffix}.txt' not in contents and f'distances_{suffix}.txt' not in contents and f'centres_{suffix}.txt' not in contents:
+        # open zip file
+        archive = zipfile.ZipFile(path)
 
-                # remove tract_lengths: get the path
-                tract_path = os.path.join(parent, f'tract_lengths_{suffix}.txt')
+        # store all newly added files
+        added = []
 
-                # remove tract_lengths
-                if os.path.exists(tract_path):
-                    os.remove(tract_path)
+        # iterate over zip content and extract only the files that are included in TO_EXRACT
+        for ext in conv.TO_EXTRACT:
+            if ext in archive.namelist():
+                # add the suffix to the newly extracted file
+                new_filename = os.path.join(parent, add_suffix(ext, suffix))
 
-                # append newly added files
-                added.append(new_filename)
+                if not os.path.exists(os.path.join(parent, ext)) and not os.path.exists(new_filename):
+                    print('old path:', os.path.join(parent, ext))
 
-    return added
+                    # extract new file
+                    archive.extract(ext, path=parent)
+
+                    # rename tract_lengths.txt to distances.txt
+                    if ext.startswith('tract_lengths'):
+                        print('ext to be replaced:', ext)
+                        new_filename = new_filename.replace('tract_lengths', 'distances')
+                        print('new path for replaced:', new_filename)
+
+                    # rename the new file
+                    os.replace(os.path.join(parent, ext), new_filename)
+
+                    # remove tract_lengths: get the path
+                    tract_path = os.path.join(parent, f'tract_lengths_{suffix}.txt')
+
+                    # remove tract_lengths
+                    if os.path.exists(tract_path):
+                        os.remove(tract_path)
+
+                    # append newly added files
+                    added.append(new_filename)
+
+        return added
 
 
 def verify_zip_rename(files):
